Changed methods to be int not double

+++ IncomeTaxStrategy.h
@@ -6,7 +6,7 @@
 class IncomeTaxStrategy : public TaxStrategy
 {
     public:
-        double calculateTax(int total) override
+        int calculateTax(int total) override
         {
             return total * (10/100);            
         };
